+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.0.6](https://github.com/dldevinc/sanic-conf/tree/v0.0.6) - 2020-11-11
+- Make `Settings` class extendable
+
+## [0.0.5](https://github.com/dldevinc/sanic-conf/tree/v0.0.5) - 2020-11-10
+- Convert settings from dict to object
+
 ## [0.0.4](https://github.com/dldevinc/sanic-conf/tree/v0.0.4) - 2020-11-03
 - Since Sanic 20.9.0 `config.from_object` is deprecated.
 - Make `settings` instance lazy.